scraper/requirements.txt
scrapling[all]>=0.4
pyyaml>=6.0
